requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.7
django-crispy-forms==1.7.2
django-leaflet==0.24.0
django-livereload-server==0.3.2
entrypoints==0.3
GeoAlchemy2==0.6.1
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.5
numpy==1.16.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.1
qtconsole==4.4.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.8
SQLAlchemy==1.3.0
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
